[not verified] Extract BarChart to its own component
