+++ projects/plugins/jetpack/_inc/client/at-a-glance/backup-upgrade/bar-chart.jsx
@@ -1,0 +1,31 @@
+/**
+ * External dependencies
+ */
+import React, { useMemo } from 'react';
+import { __ } from '@wordpress/i18n';
+import Card from 'components/card';
+
+/**
+ * Internal dependencies
+ */
+import RecordMeterBar from 'components/record-meter-bar';
+import { Popup } from './popup';
+
+const BarChart = ( { comments, plugins, posts, onClosePopup } ) => {
+	const items = useMemo( () => {
+		return [
+			{ count: posts, label: __( 'Posts', 'jetpack' ), backgroundColor: '#00BA37' },
+			{ count: plugins, label: __( 'Plugins', 'jetpack' ), backgroundColor: '#3895BA' },
+			{ count: comments, label: __( 'Comments', 'jetpack' ), backgroundColor: '#E68B28' },
+		];
+	}, [ comments, plugins, posts ] );
+
+	return (
+		<Card className="jp-dash-upgrade-backup">
+			<Popup posts={ posts } comments={ comments } onClose={ onClosePopup } />
+			<RecordMeterBar items={ items } />
+		</Card>
+	);
+};
+
+export default BarChart;
