+++ projects/plugins/jetpack/_inc/client/at-a-glance/backup-upgrade/index.jsx
@@ -1,10 +1,8 @@
 /**
  * External dependencies
  */
-import Card from 'components/card';
-import React, { useState, useCallback, useMemo } from 'react';
+import React, { useState, useCallback } from 'react';
 import { connect } from 'react-redux';
-import { __ } from '@wordpress/i18n';
 
 /**
  * Internal dependencies
@@ -15,9 +13,7 @@ import {
 	isFetchingPluginsData,
 	isFetchingStatsData,
 } from 'state/at-a-glance';
-import RecordMeterBar from 'components/record-meter-bar';
-
-import { Popup } from './popup';
+import BarChart from './bar-chart';
 
 import './style.scss';
 
@@ -26,24 +22,20 @@ const MIN_POSTS_FOR_VISIBLE_BAR = 20;
 const BackupUpgrade = ( { comments, isFetchingData, plugins, posts } ) => {
 	const [ showPopup, setShowPopup ] = useState( true );
 
-	const onClosePopup = useCallback( () => setShowPopup( false ), [] );
-
-	const items = useMemo( () => {
-		return [
-			{ count: posts, label: __( 'Posts', 'jetpack' ), backgroundColor: '#00BA37' },
-			{ count: plugins, label: __( 'Plugins', 'jetpack' ), backgroundColor: '#3895BA' },
-			{ count: comments, label: __( 'Comments', 'jetpack' ), backgroundColor: '#E68B28' },
-		];
-	}, [ comments, plugins, posts ] );
+	const onClosePopup = useCallback( () => {
+		setShowPopup( false );
+	}, [] );
 
 	return (
 		! isFetchingData &&
 		showPopup &&
 		posts > MIN_POSTS_FOR_VISIBLE_BAR && (
-			<Card className="jp-dash-upgrade-backup">
-				<Popup posts={ posts } comments={ comments } onClose={ onClosePopup } />
-				<RecordMeterBar items={ items } />
-			</Card>
+			<BarChart
+				posts={ posts }
+				comments={ comments }
+				plugins={ plugins }
+				onClosePopup={ onClosePopup }
+			/>
 		)
 	);
 };
